Cache types by cname to use for type lookups
When types are registered via the generated GtkSharp.<class>.ObjectManager,
their typeids aren't resolved yet. The type lookup eventually resolves
them as-needed, using the native type name to find the managed type.
If this mapping fails for any reason, i.e. in the cases where the
typename is not standard (like WebKitWebView, wrongly parsed as
Web.KitWebView instead of WebKit.WebView), the type lookup fails
even though the type is already registered.
Since we already have the native type name when we're generating
the RegisterType calls, we can just save it with the type and have the
type lookup use it directly.